Collapse multiple white space characters
Successive white space characters within HTML elements don't have any effect in web browsers, and so it should be safe to collapse multiple white space (including new line) charactes into a single space.

This enables the implementation of the components to be simpler, only inserting a single space between elements instead of exactly the right number of spaces and newlines.

This approach has been adopted by @natcarey at HMRC, and they've not had any issues with it so far.